Combine 3.11 and 3.12 code in the same file
