--- python-312/typing/alias_311.py
@@ -1,7 +1,0 @@
-from typing import TypeAlias, TypeVar
-
-T = TypeVar("T")
-
-Ordered: TypeAlias = list[T] | tuple[T, ...]
-
-numbers: Ordered[int] = (1, 2, 3)
