--- python-312/typing/concatenation_311.py
@@ -1,7 +1,0 @@
-from typing import TypeVar
-
-T = TypeVar("T", str, bytes)
-
-
-def concatenate(first: T, second: T) -> T:
-    return first + second
