+++ python-312/typing/generic_queue.py
@@ -1,7 +1,10 @@
 from collections import deque
+from typing import Generic, TypeVar
+
+T = TypeVar("T")
 
 
-class Queue[T]:
+class Queue(Generic[T]):
     def __init__(self) -> None:
         self.elements: deque[T] = deque()
 
@@ -10,3 +13,19 @@ class Queue[T]:
 
     def pop(self) -> T:
         return self.elements.popleft()
+
+
+# %% Python 3.12
+
+# from collections import deque
+#
+#
+# class Queue[T]:
+#     def __init__(self) -> None:
+#         self.elements: deque[T] = deque()
+#
+#     def push(self, element: T) -> None:
+#         self.elements.append(element)
+#
+#     def pop(self) -> T:
+#         return self.elements.popleft()
