--- python-312/typing/list_helpers_311.py
@@ -1,8 +1,0 @@
-from typing import TypeVar
-
-T = TypeVar("T")
-
-
-def push_and_pop(elements: list[T], element: T) -> T:
-    elements.append(element)
-    return elements.pop(0)
